# api/openapi.yaml
openapi: "3.1.0"
info:
  title: openByte Speed Test API
  version: "1.0"
  description: |
    REST API for the openByte network speed test server.
    All API endpoints are under `/api/v1/` except `/health`.
    All error responses return `{"error": "message"}`.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/download:
    get:
      summary: Download speed test (streams binary data)
      operationId: download
      tags: [SpeedTest]
      parameters:
        - name: duration
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 300
            default: 10
          description: Duration in seconds
        - name: chunk
          in: query
          schema:
            type: integer
            minimum: 65536
            maximum: 4194304
            default: 1048576
          description: Chunk size in bytes
      responses:
        "200":
          description: Binary data stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServerBusy"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/stream/start:
    post:
      summary: Start a new test stream
      operationId: startStream
      tags: [Stream]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartStreamRequest"
        description: Content-Type must be application/json. Max 1MB.
      responses:
        "201":
          description: Stream created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartStreamResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestEntityTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServerBusy"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/stream/{id}/status:
    get:
      summary: Get stream status
      operationId: getStreamStatus
      tags: [Stream]
      parameters:
        - $ref: "#/components/parameters/StreamID"
      responses:
        "200":
          description: Stream snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/stream/{id}/results:
    get:
      summary: Get stream results
      operationId: getStreamResults
      tags: [Stream]
      parameters:
        - $ref: "#/components/parameters/StreamID"
      responses:
        "200":
          description: Completed stream results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamSnapshot"
        "202":
          description: Stream not yet completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: stream not completed
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/stream/{id}/cancel:
    post:
      summary: Cancel a stream
      operationId: cancelStream
      tags: [Stream]
      parameters:
        - $ref: "#/components/parameters/StreamID"
      responses:
        "200":
          description: Stream cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: cancelled
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/stream/{id}/complete:
    post:
      summary: Complete a stream (client mode)
      operationId: completeStream
      tags: [Stream]
      parameters:
        - $ref: "#/components/parameters/StreamID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, metrics]
              properties:
                status:
                  type: string
                  enum: [completed, failed]
                metrics:
                  $ref: "#/components/schemas/Metrics"
        description: Content-Type must be application/json. Max 1MB.
      responses:
        "200":
          description: Stream completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "413":
          $ref: "#/components/responses/RequestEntityTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    StreamID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Stream UUID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MethodNotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict (e.g. stream already exists)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RequestEntityTooLarge:
      description: Request body exceeds size limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnsupportedMediaType:
      description: Content-Type must be application/json
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerBusy:
      description: Server at capacity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    StartStreamRequest:
      type: object
      required: [protocol, direction]
      properties:
        protocol:
          type: string
          enum: [tcp, udp]
        direction:
          type: string
          enum: [download, upload, bidirectional]
        duration:
          type: integer
          minimum: 1
          maximum: 300
          default: 30
        streams:
          type: integer
          minimum: 1
          maximum: 32
          default: 4
          description: Max 32 (configurable). Default 4.
        packet_size:
          type: integer
          minimum: 64
          maximum: 9000
          default: 1400
        mode:
          type: string
          enum: [client, proxy]
          default: proxy
    StartStreamResponse:
      type: object
      properties:
        stream_id:
          type: string
          format: uuid
        websocket_url:
          type: string
        test_server_tcp:
          type: string
        test_server_udp:
          type: string
        status:
          type: string
        mode:
          type: string
          enum: [client, proxy]
    LatencyMetrics:
      type: object
      properties:
        min_ms:
          type: number
          format: double
        max_ms:
          type: number
          format: double
        avg_ms:
          type: number
          format: double
        p50_ms:
          type: number
          format: double
        p95_ms:
          type: number
          format: double
        p99_ms:
          type: number
          format: double
        count:
          type: integer
    RTTMetrics:
      type: object
      properties:
        baseline_ms:
          type: number
          format: double
        current_ms:
          type: number
          format: double
        min_ms:
          type: number
          format: double
        max_ms:
          type: number
          format: double
        avg_ms:
          type: number
          format: double
        jitter_ms:
          type: number
          format: double
        samples:
          type: integer
    Metrics:
      type: object
      properties:
        throughput_mbps:
          type: number
          format: double
        throughput_avg_mbps:
          type: number
          format: double
        latency_ms:
          $ref: "#/components/schemas/LatencyMetrics"
        rtt:
          $ref: "#/components/schemas/RTTMetrics"
        jitter_ms:
          type: number
          format: double
        packet_loss_percent:
          type: number
          format: double
        bytes_transferred:
          type: integer
          format: int64
        packets_sent:
          type: integer
          format: int64
        packets_received:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        stream_count:
          type: integer
    NetworkInfo:
      type: object
      properties:
        client_ip:
          type: string
        server_ip:
          type: string
        isp:
          type: string
        asn:
          type: integer
        ipv6:
          type: boolean
        nat_detected:
          type: boolean
        mtu:
          type: integer
        hostname:
          type: string
    StreamSnapshot:
      type: object
      properties:
        config:
          type: object
          properties:
            id:
              type: string
              format: uuid
            protocol:
              type: string
            direction:
              type: string
            duration:
              type: integer
            streams:
              type: integer
            packet_size:
              type: integer
            client_ip:
              type: string
        status:
          type: string
          enum: [pending, starting, running, completed, failed, cancelled]
        progress:
          type: number
          format: double
        metrics:
          $ref: "#/components/schemas/Metrics"
        network:
          $ref: "#/components/schemas/NetworkInfo"
        error:
          type: string
          description: Error message when status is failed
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
...